+++ packages/fast-components-styles-msft/src/index.ts
@@ -6,6 +6,7 @@ export { ButtonStyles };
 
 import CallToActionStyles from "./call-to-action";
 export { CallToActionStyles };
+export * from "./call-to-action";
 
 import CaptionStyles from "./caption";
 export { CaptionStyles };
