+++ packages/fast-glyphs-msft/.npmignore
@@ -1,0 +1,5 @@
+# Source files
+src/
+
+# babel config
+.babelrc
